Config Play Song Screen part. 62.

+++ app/src/main/java/com/example/musicapp/presentation/play_song/widgets/PlaySongContent.kt
@@ -23,7 +23,7 @@ fun PlaySongContent(
     duration: Long,
     isPlaying: Boolean,
     isBuffering: Boolean,
-    onSeekChange: (newValue: Float) -> Unit,
+    onSeekChange: (newValue: Long) -> Unit,
     onPlayPauseToggle: () -> Unit
 ) {
     Column(
